Refactor 17c cryoscope plotting and consolidate FIR utilities.
Move auto-tuned FIR math into common fir_utils, keep all matplotlib in
plotting.py, and gate diagnostic figures behind debug_plots with Debug:
suptitles on the routine flux-response plot only when debug is off.
