+++ qualibration_graphs/superconducting/calibration_utils/qubit_flux_short_distortion/parameters.py
@@ -64,7 +64,7 @@ class NodeSpecificParameters(RunnableParameters):
     use_fir: bool = False
     """Run FIR analysis after IIR. Default False."""
     fir_max_taps: int = 48
-    """Forward / inverse FIR length (fixed; no auto length search)."""
+    """Upper bound on forward/inverse FIR tap count for auto L selection."""
 
     update_state: bool = False
     """Master gate for writing fitted filters into QUAM state."""
@@ -78,8 +78,10 @@ class NodeSpecificParameters(RunnableParameters):
     freq_to_flux_source: Literal["auto", "ramsey", "spectroscopy", "quad_term"] = "auto"
     """Which frequency->voltage relation to use when picking the cryoscope Z amplitude and when converting the measured f(t) into a flux step response. 'auto' (default) tries Ramsey vs flux (09a), then qubit spectroscopy vs flux (03b), then the quadratic freq_vs_flux_01_quad_term; the other values force one specific source. Run IDs are never entered by hand: they are read from each qubit's extras ('ramsey_vs_flux_calibration_load_id' for 09a, 'qubit_spectroscopy_vs_flux_load_id' for 03b), so run those nodes with save_load_id=True first."""
     debug_plots: bool = False
-    """If True, also show cryoscope frequency, unwrapped phase, freq-vs-flux curve,
-    and FIR resampling check (each on a ``QubitGrid``)."""
+    """If True, show diagnostic figures: cryoscope frequency/phase/freq-vs-flux,
+    FIR resampled grid, raw I vs frame, and all FIR diagnostic plots."""
+    log_time_axis: bool = False
+    """If True, plot flux response (and debug cryoscope frequency) vs log time; else linear."""
 
 
 class Parameters(
